apps/api/pyproject.toml
[project]
name = "abenix-api"
version = "0.1.0"
description = "Abenix API server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.2.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "stripe>=9.0.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "croniter>=2.0.0",
    "boto3>=1.34.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    # Meeting Representative — LiveKit server SDK for token minting, and
    # pypdf for persona file ingestion.
    "livekit-api>=0.7.0",
    "pypdf>=4.0.0",
    "openai>=1.30.0",
    "pinecone>=5.0.0",
    # BPM Analyzer — page rasterization for vision-based diagram input
    "PyMuPDF>=1.24.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.0",
    # Wave-2 scaling: API publishes to NATS when scaling.execRemote=true
    # and queueBackend=nats. Kept optional in the selector so a missing
    # client falls back to Celery without breaking the request.
    "nats-py>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
